backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
mysql-replication
websockets
python-dotenv
pydantic
pydantic-settings
httpx
aiofiles
sshtunnel
paramiko
python-jose[cryptography]
passlib[bcrypt]
paho-mqtt
python-multipart
firebase-admin
